Serialize and deserialize world save data (snapshots and entity collections) as a JSON tree (#11939)
Replaces our "`BlobTree`" + RON with a pure-JSON based document format
for saving (pieces of) worlds, aka "world slices". This is used both for
saving worlds and for saving persisted collections (used in prefabs,
inventory, etc).

The "blob tree" is replaced by a "`JsonTree`", which inherits most of
the properties of the blob tree but is itself valid JSON and the
payloads carried must also be valid JSON. See new `jsontree.rs` for more
information.

The main structure is a tree of nodes, where data is laid out in a
"struct of arrays" fashion. First comes a list of all the entities in
the world slice, along with a list of the GUIDs of any behaviors
attached to said entities, in the order those behaviors were added. Then
comes the behavior stores, each behavior separated by GUID, then the
data for any entity's individual behavior in a sub-node under the store.
The aspects are the same but under an aspects node.

As part of this change, several content store prefixes have been
incremented, providing a clean slate for world meta, snapshots,
inventory, persisted collections, prefabs, user info, and activity
progression.

One key difficulty that resulted in not attempting to migrate data
forward is that the serialization format for several types containing
maps needed to be updated in a breaking way, as JSON-native maps only
support string keys. Thus the default serialization strategy for maps
going forward is serializing as a sequence of `(key, value)` pairs
rather than serde-native Maps.

---------

Co-authored-by: Embark Bot <47564659+embark-bot@users.noreply.github.com>